Add Annotations to chart
Use third-party plugin to add annotations to the chart.
Just display a single note based on static hard-coded annotation object first.

Add a click event to set the x and y position when the note is dragged.
NB this is written into the text filed as a way of passing the data out of the chart object