set  y axis label rounding, fix scatter tooltip ------------------------------------------------------------------------ use the decimal point to round the y axis figures With scatter charts clear the main chart tooltip and use a custom one
Why is this change necessary?
*

How does it address the issue?

A good commit message briefly summarises the "what" for scanning purposes, but also includes the "why". If the "what" in the message isn't enough, the diff is there as a fallback. This isn't true for the "why" of a change - this can be much harder or impossible to reconstruct, but is often of great significance.

*

What side effects does this change have?
*
Further paragraphs come after blank lines.

- Bullet points are okay, too